+++ src/checklist.cpp
@@ -42,7 +42,7 @@ static int ret = 1;
 
 static void close_cb(Fl_Widget *, long p) {
   win->hide();
-  ret = (int) p;
+  ret = p;
 }
 
 int dialog_checklist(std::string checklist_options, bool return_value, bool check_all, char separator)
